Feat: Review CRUD 기능 구현,

+++ src/main/java/kea/dpang/item/dto/ReadCartItemDto.java
@@ -6,7 +6,7 @@ public class ReadCartItemDto {
 
     private final String thumbnailImage; // 상품 이미지 URL
     private final String itemName; // 상품 이름
-    private final Long discountPrice; // 상품 가격
+    private final int discountPrice; // 상품 가격
 
     public ReadCartItemDto(Item item) {
         this.thumbnailImage = item.getThumbnailImage();
